Ecommerce-api/requirements.txt
django
django-cors-headers
djangorestframework-simplejwt
drf_yasg